Fix wave restart for cold start
The stage job was incorrectly putting wave restarts into the gfs
directory. The forecast job looks for them in the gdas directory,
so this is updated.

Additionally, the restarts were also not being copied from the
staged directory to `$DATA`, so now they are. The process is
identical to that of non-RERUN warm starts, so the code is re-
factored a bit to avoid duplication.

Resolves #3109