Update stage and archive to conditionally handle ca_data
The stage and archive jobs were assuming that the ca_data restart
files will always be present. However, these are only produced
when cellular automata is run. So now they are conditionally
handled based on `DO_CA`.

To accomodate this, the `DO_CA` setting had to be moved out of
`config.fcst`/`config.efcs` and into `config.base`.