Upate compression options for GEFS history files
Different compression options are applied for the high resolution
history files. The value of quantize_nsd will be different
depending on the value of quantize_mode, so a fix is required
to obtain the same behavior as the previous compression options.
This option needs to be updated for both GFS and GEFS.

Resolves: #3178