Update for syndat_getjtbul code
The gfs_atmos_tropcy_qc_reloc did not properly parse the vitals info from a raw bulletin
with a tab in the storm name field for storm "SIX". The code sorc/syndat_getjtbul.fd/getjtbul.d
converting a tab to a space worked well for storms that have at least 4 characters in their names,
but does not work as well for storms with 3-character names. This code change will fix the
parsing problem for 3-character storms.

Refs #2591