Fix for syndat_getjtbul code parsing 3-character storm names (GFSv16.3.21) (#3200)
Update the `sorc/syndat_getjtbul.fd/getjtbul.f` code to resolve a problem parsing 3-character storm names.
This update is being included in the GFSv16.3.21 package (with approval by NCO).

From the developer @Qingfu-Liu:

> The `gfs_atmos_tropcy_qc_reloc` job did not properly parse the vitals
info from a raw bulletin with a tab in the storm name field for storm
"SIX". The code sorc/syndat_getjtbul.fd/getjtbul.f converting a tab to a
space worked well for storms that have at least 4 characters in their
names, but does not work as well for storms with 3-character names. This
code change will fix the parsing problem for 3-character storms.

Refs #2591