Add sfs as valid system
Adds sfs as a valid option for NET.

To start, the GEFS system is generally just copied wholesale for SFS.
This includes the extract_vars job.

Other than base and resources, config files link to the GEFS versions,
just as GEFS config files point to the GFS versions except where they
have needed to be changed.

The temporarily SFS_POST option has been removed.

The sfs test case is moved from a separate sfs directory to the pr
directory, but ICs are still needed before it is functional.

Resolves #2271