Modify link_workflow to try to address CI failures
Changes were made to sorc/link_workflow.sh and
ush/parsing_namelists_FV3.sh to try and address the CI
failures. There were still some discrepancies in the
namelists where XR_CNVCLD was being used but was not
declared in the template and INCREMENT_FILE_ON_NATIVE_GRID
was in the template but the default value might have been
incompatible with some of the CI cases.

If the CI tests pass, then a more permanent solution should be
found and the code will need to be cleaned up before the
final merge into the GW repo.