Update boundary point, wave gempak jobs for file rename
Updates the boundary point j-jobs and the gempak scripts for the
wave file rename.

The gempak scripts are also refactored to meet style guidelines. As
part of this, the job is changed to run one forecast hour at a time,
with a rocoto metatask for development. Grouping is done for this
metatask as with other product metatasks to reduce the job count.
For dependency reasons, these settings must be the same as the wave
gridded post, so the max tasks and fhrs are read from that config
set in rocoto.

Like the regular point job, the boundary point j-jobs are not
refactored as the point job will be redone soon anyway. Only changes
necessary for the new filenames are made.