Update operational workflow to work with GSI develop branch (#3068)
Update operational workflow to work with GSI develop branch. This mostly
involves changing the links to elements that have been moved out of the
GSI repository since GFSv16.3. The ultimate aim is to be able to cycle
with v16 workflow but using a version of GSI close to develop. This will
not reproduce operations.

Refs #3045