Turn off soil moisture relaxation. (#4353)
This PR corrects the
soil relaxation bug in v17 retros, resolving issue
[#](https://github.com/NOAA-EMC/global-workflow/issues/4326)

In develop, this fix was included in a PR
[#4327](https://github.com/NOAA-EMC/global-workflow/pull/4327) which
included an additional change to correct a long-standing error in the
time input to global_cycle. The latter change has not been tested
sufficiently for inclusion into v17 retros, so I've created this PR to
merge in only the soil relaxation change to v17.